Branch update

See merge request c2038058/ase_y2s1_cp_g4!44

+++ src/main/java/com/example/clientproject/data/shops/Shops.java
@@ -25,7 +25,7 @@ import java.util.List;
 public class Shops {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int shopId;
+    private long shopId;
     private String shopName;
     private String shopDescription;
     private String shopWebsite;
